project validated in Moulinette

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: jusato <jusato@student.42sp.org.br>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/05 20:13:34 by jusato            #+#    #+#             */
-/*   Updated: 2022/04/05 22:32:03 by jusato           ###   ########.fr       */
+/*   Updated: 2022/05/09 04:33:48 by jusato           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (*str)
